video_tool/requirements.txt
moviepy
pillow
numpy
ddgs
requests